Extracted Tags to Core.Tags
