+++ Api/WebScrap.API/Extract.cs
@@ -1,7 +1,6 @@
 using System.Collections.Immutable;
-using WebScrap.Tags;
+using WebScrap.Core.Tags;
 using WebScrap.Css;
-using WebScrap.Common.Contracts;
 using WebScrap.Css.Preprocessing;
 using WebScrap.Common;
 
@@ -20,22 +19,29 @@ public static class Extract
         ReadOnlySpan<char> css)
     {
         html = html.TrimStart(' ');
-        var cssTokens = PreprocessingAPI.Process(css);
-        var cssProcessor = new CssProcessor(cssTokens);
-        cssProcessor.Run(html);
-        var tagIndexes = cssProcessor.TagIndexes;
-        var tagRanges = ExtractTagRanges(cssProcessor, html.ToString(), tagIndexes);
+        var tagIndexes = GetTagIndexes(css, html);
+        var tagRanges = ExtractTagRanges(html.ToString(), tagIndexes);
         var tagStrings = ExtractStrings(html.ToString(), tagRanges);
         return tagStrings;
     }
 
+    private static List<int> GetTagIndexes(
+        ReadOnlySpan<char> css, 
+        ReadOnlySpan<char> html)
+    {
+        var cssTokens = PreprocessingAPI.Process(css);
+        var cssProcessor = new CssProcessor(cssTokens);
+        cssProcessor.Run(html);
+        return cssProcessor.TagIndexes;
+    }
+
     private static ImmutableArray<Range> ExtractTagRanges(
-        ProcessorBase processor,
         string html,
         IEnumerable<int> tagIndexes)
         => tagIndexes.Select(tagIndex =>
         {
             var substring = html[tagIndex..];
+            var processor = new TagsProcessor();
             processor.Run(substring);
             var offset = processor.CharsProcessed;
             return tagIndex..(tagIndex + offset);
