+++ Api/WebScrap.API/WebScrap.API.csproj
@@ -2,7 +2,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\Css\WebScrap.Css\WebScrap.Css.csproj" />
-    <ProjectReference Include="..\..\Tags\WebScrap.Tags\WebScrap.Tags.csproj" />
+    <ProjectReference Include="..\..\Core.Tags\WebScrap.Core.Tags\WebScrap.Core.Tags.csproj" />
     <ProjectReference Include="..\..\Html\WebScrap.Html.Extracting\WebScrap.Html.Extracting.csproj" />
   </ItemGroup>
