+++ Core.Tags/WebScrap.Core.Tags.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebScrap.Tags.Tests", "WebScrap.Tags.Tests\WebScrap.Tags.Tests.csproj", "{2EF0FD35-7C6C-4CE7-8BA6-BE00FB7402FA}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebScrap.Core.Tags.Tests", "WebScrap.Core.Tags.Tests\WebScrap.Core.Tags.Tests.csproj", "{2EF0FD35-7C6C-4CE7-8BA6-BE00FB7402FA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebScrap.Tags", "WebScrap.Tags\WebScrap.Tags.csproj", "{24E53336-6607-4660-B194-297813769605}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebScrap.Core.Tags", "WebScrap.Core.Tags\WebScrap.Core.Tags.csproj", "{24E53336-6607-4660-B194-297813769605}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
