+++ Core.Tags/WebScrap.Core.Tags/Usings.cs
@@ -1,4 +1,4 @@
 using System.Runtime.CompilerServices;
 
-[assembly: InternalsVisibleTo("WebScrap.Tags.Tests")]
+[assembly: InternalsVisibleTo("WebScrap.Core.Tags.Tests")]
 [assembly:InternalsVisibleTo("WebScrap.Tests.IntegrationTests")]
