+++ Css/Tests/WebScrap.Css.Tests/WebScrap.Css.Tests.csproj
@@ -21,7 +21,7 @@
   <ItemGroup>
     <ProjectReference Include="..\..\WebScrap.Css\WebScrap.Css.csproj" />
     <!-- TagsFactory -->
-    <ProjectReference Include="..\..\..\Tags\WebScrap.Tags\WebScrap.Tags.csproj" />
+    <ProjectReference Include="..\..\..\Tags\WebScrap.Core.Tags\WebScrap.Core.Tags.csproj" />
   </ItemGroup>
 
 </Project>
