+++ Css/WebScrap.Css/WebScrap.Css.csproj
@@ -10,6 +10,7 @@
   <ItemGroup>
     <ProjectReference Include="..\WebScrap.Css.Preprocessing\WebScrap.Css.Preprocessing.csproj" />
     <ProjectReference Include="..\WebScrap.Css.Matching\WebScrap.Css.Matching.csproj" />
+    <ProjectReference Include="..\..\Core.Tags\WebScrap.Core.Tags\WebScrap.Core.Tags.csproj" />
   </ItemGroup>
 
 </Project>
