+++ Html/WebScrap.Html.Extracting/ExtractingAPI.cs
@@ -1,5 +1,4 @@
 ﻿// using System.Collections.Immutable;
-// using WebScrap.Common.Contracts;
 // using WebScrap.Common.Css;
 // using WebScrap.Html.Extracting.Listeners;
