--- WebScrap.Common/Contracts/TagFactoryBase.cs
@@ -1,20 +1,0 @@
-using WebScrap.Common.Tools;
-using WebScrap.Common.Tags;
-
-namespace WebScrap.Common.Contracts;
-
-public abstract class TagFactoryBase
-{
-    public TagBase CreateTagBase(ReadOnlySpan<char> tag)
-    {
-        var tagCreator = GetTagCreator(tag);
-        tag = tag.Clip("<", ">");
-        return tag.IndexOf('/') switch 
-        {
-            0 => tagCreator.Create(tag[1..]),
-            _ => tagCreator.Create(tag),
-        };
-    }
-
-    protected abstract ITagCreator GetTagCreator(ReadOnlySpan<char> tag);
-}
